requirements.txt
setuptools
xpander-sdk
pydantic
loguru
httpx
httpx_sse
python-dotenv
pytest
black
pre-commit
pytest-asyncio
sqlalchemy
agno
psycopg2-binary
nest-asyncio
